year_I/VRNN/static_gestures_recognition/requirements.txt
tensorflow == 2.0
tensorflow-gpu == 2.0
numpy == 1.16.4
Pillow >= 7.00, < 8
scipy >= 1.4, < 2
torchvision == 0.4.2
matplotlib == 3.1.2
scikit-image == 0.16.2
scikit-learn == 0.22.1
gast == 0.2.2
tqdm == 4.41.1
